fixing protocol to get svpncookie
some change happened over last weeks, so that the /remote/logincheck
responded to POST http requests with 405 messages and did not include
svpncoookie. changes to the agent in the request (away from SV1)
resolved the 405, but still didnt include the cookie and instead
contained some sslvpn_login_permission_denied message. so, thats
was a dead end.
the only workaround i could find was to switch to GET instead of
the POST request.